Create doc
Before we discuss artificial intelligence (AI), it is necessary to define the term. We 
can begin by defining what we mean by intelligence. One idea put forward by Simmons and 
Chappell states that intelligence is an overt ability to solve specific problems and an innate 
ability to learn solutions for new problems. Considering this, we can define AI to be an 
artificial entity capable of solving problems and learning solutions for new problems. This 
implicitly requires the entity to be able to perceive its environment, (i.e. detect input data and 
its parameters), search and perform pattern recognition (i.e. identify and recognize features of 
the problem), plan and execute an appropriate course of action and perform inductive 
reasoning to derive general principles (i.e. learn from experience). In the context of this 
paper, AI is hardware or software designed to carry out a specific task while possessing the 
above characteristics. The applied science of AI seeks to improve computer systems until 
their function is equal to, or greater than, that of a human performing the same task.